Prevent corrupting the PG DB during initialization
If the bootstrap script fails, Make will delete the file it was trying
to 'build'. In this case that meant deleted PG_VERSION. By make a
directory, postgresql/data, instead of a file the target we prevent
this deleting behavior.

We want this because the bootstrap can take forever while it generates
seed data and the user may press Ctrl-C.