Save some variables in .env.mk for customization
To make it easier to customize variables, add the use of `.env.mk`.

This is what happens:

The file `.env.mk` is included optionally (hence the leading dash). So
when the file exists, it is included, and the variables are loaded.

The variables that might be overridden by the above file, are
optionally assigned with `?=`. So only when they weren't set already,
they are set to the defaults.

An extra target `.env.mk` is added. This depends on all the included
makefiles (i.e. for now only `Makefile`). So when that file changes,
`.env.mk` is rebuild.

Building `.env.mk` consists of looping through all the variable names
mentioned in `env_variables` (only `postgres_bin_dir` at the moment)
and writing the assignment expression in that file. So the next run,
the Makefile can include that and will have the values already set.

Now here is the cool thing. Users can customize the values in
`.env.mk`. So when you intentionally want to use the postgreSQL 10.x
`bin` dir, you can set it in `.env.mk`. And because the file is
included and the variables are optionally assigned, rebuilding the
file will retain the changes made by the user.