Updated Rubocop definitions to work with 0.59.x
There is also a workaround to enable rubocop support integration on
a variety of code editors. As `gdk-rubocop.yml` is not a standard
file, many editors will just ignore it and run rubocop with defaults.

Because we only have ruby code in a few folders, we can add a
`.rubocop.yml` file on these folders and inherit from `gdk-rubocop.yml`.

We can't just rename `gdk-rubocop.yml` as this would interfere with the
configuration file inside the `gitlab` folder.