Check for configuration updates
This change will prompt GDK users when the underlying GDK config
template changes.

This approach is already used for unicorn.rb and puma.rb files, and
this change extends the concept to all configuration files in GDK.

This means that we a change is made to a configuration file in GDK, the
local file will be overwritten with the change, except if the user has
made changes locally, in which case the user will receive a notification
stating that their local configuration is out of date.