Validate whitespace
This change adds a `.editorconfig` for GDK, to help keep whitespace
under control. It also adds `make verify` and `make fix` tasks to
verify and fix.

An additional CI task, for ensuring that the rules are in-compliance
has been added.

The ruleset is designed not to impact projects hosted in
subdirectories, which is how GDK is designed to operate.