Writing of the gitlab.yml uses Ruby now
Prior to this change, `ed` was used and the syntax and language are very
complete, but hard to understand what's going on. This limits the
ability to contribute. Also more involved logic is now hard to get
rigth and understand/maintain later.

This change does have the side effect that setting the correct value in
GitLabs `config/gitlab.yml.example` does not propegate by default to
everyones machine.