Provide more feedback in support/wait-postgresql-ready
We had multiple team members see their `gdk install` fail because psql
wasn't in the PATH or needed to be reinstalled due to readline updates
in their environment. The problem wasn't obvious because
wait-postgresql-ready quietly failed without indicating a reason.

To make this clearer, use `Popen.open3` instead of `system`. We now:

1. Fail hard with an exception if the psql binary doesn't exist
2. Show the last error message to give more feedback