Convert defaults file into ruby code
Instead of crazy tricks with ERB in YAML, define the config directly
in ruby.

With method_missing we're dynamically generating methods of each of
the settings. These settings follow the same structure of the YAML
file people can use to override their settings.

So the settings are applied as following:
- from gdk.yml
- from lose files like: hostname, port, webpack_port, etc.
- from defaults defined in lib/gdk/config.rb directly