Rsync is a required dependency for Gitaly now
When running for example `make update`, Gitaly will require rsync to
speed up the operation of copying files. This dependency wasn't documented
in the GDK yet, but now is.

For some OS's I've not filled in the dependency, as I couldn't find what rsync
was called for their package manager.

Related to: https://gitlab.com/gitlab-org/gitaly/merge_requests/1272