Move Gitaly config to Rake tasks
While not necessarily required, moving all Gitaly tasks to Rake allowed
for easier access to them and allowed easier configuration of Gitaly and
GitLab alike.

This flexibility is leveraged for the new rake tasks for `praefect`.
`rake praefect:enable` and `rake praefect:disable` are introduced to use
Praefect as a proxy in the GDK and allow for later iterations which will
run a cluster of Gitaly servers.

Supersedes: https://gitlab.com/gitlab-org/gitlab-development-kit/merge_requests/645