Add a rake task to make git config recommendations
Advising the GitLab community to set certain config flags for quality of
life recommendations is hard to achieve as it scales with the number of
people that know about them. When the knowledge is put into a rake task
the only knowledge to spread is; run `rake git:configure`.

Hopefully this will allow people to be more effective without spending a
lot of time to get sane value in their Git configuration.