Run https on a custom domain
This contains a bunch of changes to be able to run GDK on a custom
domain (using /etc/hosts).

You should add a loopback interface and point a `.localhost` domain to
that IP in /etc/hosts. Configure it in `gdk.yml` and you should have a
fully working GDK with a domain and https.

I've been using `mkcert` to generate a certificate for localhost.