Add tag.sort config recommendation
Git, by default, sorts tags lexicographicly. Meaning that if a
repository has two tags, v1.0 and v1.1, the list is:

v1.0
v1.1

On repositories with a lot of tags,, like GitLab has a lot of, this
makes the plain `git tag` command less useful as the first releases are
listed first, while the user is probably more interested in current
releases. By reverse lexicographicly sorting the tags, this is solved.