Fix bundler detection
When multiple versions of bundler are available, bundle will default to
the version specified in Gemfile.lock.

gitlab-development-kit has no Gemfile.lock and this will result in
invoking the latest version available.

However, it is possible to invoke bundle with a specific version, and it
will fail if it's not available.

This patch will change GDK::Dependencies::Checker#check_bundler_version
to invoke bundle _required_version_ --version and check the return value