jaeger: fix download on musl-based systems
In order to set up jaeger, we check the OSTYPE variable and choose the
correct binary archive to download. We currently support both
glibc-based Linux systems as well as macOS, but this leaves out some
more exotic platforms like musl-based Linux.

Recognize "linux-musl" and treat it the same as "linux-gnu". The
binaries contained in the archive are statically linked anyway and thus
do not depend on any external libc, making them work on both glibc- and
musl-based systems.