doc: storage: Provide instructions to use Linux' device mapper
The current instructions for simulating delays for storage depends on
the use of sshfs. One obvious benefit is that it's cross platform and
doesn't require any kernel support except for FUSE, but on the other
hand it's not all that useful due to not being as flexible as other
solutions.

On Linux, we can instead use some device mapper targets, most
importantly the targets "delay", "dust" and "flaky". These allow you to
exactly tweak several parameters like read/write delays, bad disk
sectors and the likelihood of I/O operations causing failures. This
flexibility is not possible with sshfs.

Document the setup of these device mapper targets to help others
discover this superior, but more complex, option.