openssh: Accept GIT_PROTOCOL environment variable
Git's SSH transport sets up the GIT_PROTOCOL environment variable in
order to communicate to the server which protocols it supports. As sshd
will not accept any environment variables by default, we need to
explicitly accept GIT_PROTOCOL to make the integration work.

Add the required statement to the sshd_config template to make this work
as expected.