Fix Gitaly address for tests
Previously only the Gitaly process was started in tests, but the Gitaly
address was configured to use `praefect.socket` because the GDK may have
been configured to use Praefect. However, the change in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26323 caused specs
to fail because the test environment expects both a Gitaly and Praefect
address to be available on distinct sockets
(e.g. `tmp/tests/gitaly.socket` and `tmp/tests/praefect.socket`). Since
`gitaly.socket` wasn't available, the test would fail to start.

Now, the GDK should always be using `gitaly.socket` for tests, so we
need to adjust this config accordingly.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/212509