Improve automation for configuring docker registry
Configuration is now driven through gdk.yml

In particular, it adds support for configuring a local self-signed
registry, which is closer to how a production GitLab installation would
work, and more secure than a passwordless or unsigned registry.

The updated documentation outlines the new features and shows how a
local registry can be configured in a way that does not require editing
vendored CI templates when working on Auto DevOps (otherwise
--insecure-registry modifications are required to GitLab source code
during development).