Disable EE load balancing when Geo is in use
When Geo is in use, we can't use database load balancing on the
Geo secondary because load balancing attempts to query the write
location on the primary database:

```
PG::ObjectNotInPrerequisiteState at /admin
ERROR:  recovery is in progress
HINT:  WAL control functions cannot be executed during recovery.
```

Ideally we'd only disable this on the Geo secondary and keep it enabled
on the Geo primary, but as a first step we can just disable this
outright for development purposes.