Use ASA silent fail for Auto DevOps SSH config
The existing config automatically opens a web browser when the GDK
attempts to make an SSH request and the device is not currently signed
into ASA. If a GDK is left running over night this can result in
thousands of browser tabs opening and an incident being created for the
ASA team as a result of the unusual authentication activity.

This commit updates the docs to instruct developers to use a "silent
fail" config that will quitely kill unauthenticated requests. Developers
will have to manually sign in from the command line when their ASA
sessions expire (by default, every 9 hours).