Remove empty gitlab-workhorse/config.toml.example
The GDK runs `touch gitlab-workhorse/config.toml.example`, but that file
was introduced in v8.34.0. Running `make gitlab-workhorse-update` now
fails for existing installations because this version would overwrite a
file that existed in the directory.

To prevent this error, this change causes `make gitlab-workhorse-update`
to remove this file if is empty.